# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CUSIT University Event Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/registrations:
    get:
      operationId: listEventRegistrations
      tags: [registrations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Registrations
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Registration" }
  /events/{id}/register:
    post:
      operationId: registerForEvent
      tags: [registrations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Registration" }
  /events/{id}/unregister:
    post:
      operationId: unregisterFromEvent
      tags: [registrations]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Unregistered }
...
  /me/notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Marked }
  /events/{id}/feedback:
    get:
      operationId: listEventFeedback
      tags: [feedback]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Feedback
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Feedback" }
    post:
      operationId: submitFeedback
      tags: [feedback]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FeedbackPayload" }
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Feedback" }
...
  /reports/registrations-trend:
    get:
      operationId: getRegistrationsTrend
      tags: [reports]
      summary: Registrations grouped by day for the last N days
      parameters:
        - { name: days, in: query, schema: { type: integer, default: 14 } }
      responses:
        "200":
          description: Trend data
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TrendPoint" }
  /reports/top-events:
    get:
      operationId: getTopEvents
      tags: [reports]
      summary: Top events by registrations
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 5 } }
      responses:
        "200":
          description: Top events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TopEvent" }
  /reports/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [reports]
      summary: Recent activity feed across the system
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: Activity items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ActivityItem" }
  /reports/event/{id}:
    get:
      operationId: getEventReport
      tags: [reports]
      summary: Per-event report (registrations, attendance, certs, feedback avg)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Event report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventReport" }
components:
  schemas:
...
    UserRole:
      type: string
      enum: [student, faculty, coordinator, society_head, admin]
    EventStatus:
      type: string
      enum: [pending, approved, rejected, completed, cancelled]
    EventCategory:
      type: string
      enum: [seminar, workshop, competition, sports, society, cultural, technical]
...
    Event:
      type: object
      required: [id, title, description, category, status, organizerId, organizerName, startsAt, endsAt, venue, capacity, registrationCount]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category: { $ref: "#/components/schemas/EventCategory" }
        status: { $ref: "#/components/schemas/EventStatus" }
        organizerId: { type: integer }
        organizerName: { type: string }
        organizerRole: { $ref: "#/components/schemas/UserRole" }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        venue: { type: string }
        capacity: { type: integer }
        registrationCount: { type: integer }
        bannerUrl: { type: string, nullable: true }
        rejectionReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...
    Registration:
      type: object
      required: [id, eventId, userId, userName, registeredAt]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        userEmail: { type: string }
        userDepartment: { type: string, nullable: true }
        registeredAt: { type: string, format: date-time }
...
    Feedback:
      type: object
      required: [id, eventId, userId, userName, rating, createdAt]
      properties:
        id: { type: integer }
        eventId: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    FeedbackPayload:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, nullable: true }
...
    TrendPoint:
      type: object
      required: [date, count]
      properties:
        date: { type: string }
        count: { type: integer }
    TopEvent:
      type: object
      required: [eventId, title, registrationCount, attendanceRate]
      properties:
        eventId: { type: integer }
        title: { type: string }
        category: { $ref: "#/components/schemas/EventCategory" }
        registrationCount: { type: integer }
        attendanceRate: { type: number }
    ActivityItem:
      type: object
      required: [type, message, timestamp]
      properties:
        type: { type: string, enum: [event_created, event_approved, event_rejected, registration, certificate, feedback] }
        message: { type: string }
        eventId: { type: integer, nullable: true }
        timestamp: { type: string, format: date-time }
    EventReport:
      type: object
      required: [event, registrationCount, attendedCount, absentCount, certificatesIssued, avgRating, feedbackCount]
      properties:
        event: { $ref: "#/components/schemas/Event" }
        registrationCount: { type: integer }
        attendedCount: { type: integer }
        absentCount: { type: integer }
        certificatesIssued: { type: integer }
        avgRating: { type: number }
        feedbackCount: { type: integer }