# lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CUSIT University Event Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: number;
  eventId: number;
  userId: number;
  userName: string;
  /**
   * @minimum 1
   * @maximum 5
   */
  rating: number;
  comment?: string | null;
  createdAt: Date;
}